made usage of α/a consistent
Also when running it locally I get an `┌ Warning: dt <= dtmin. Aborting. There is either an error in your model specification or the true solution is unstable.`. The solution would be only one data point. I am not quite sure how to fix that. Randomly changing stuff in line 76 makes results in more data points but not the in the plot from the tutorial.