Refactor Alias Elimination some more
This is a follow up to #1301 to finish refactoring the alias elimination
code. It mostly puts the result of AliasElimination into an `AliasGraph`
datastructure, but while we're at it, it also improves the asymptotics
of alias elimination slightly (as noted possible in #1301). There's further
asymptotics improvement possible that I've marked in the code, but don't
intend to pursue for the moment unless alias elimination starts showing
up in benchmarks. This also fixes a number of small bugs: For example,
the iteration to convergence in the back substitution was previously
an accidental infinite loop. Additionally, it should be fully able to
handle integer coefficients other than (1, -1), though this capability
is not currently enabled.