Replace open coded tarjan algorithm by Graphs.jl version
Besides avoiding redundancy and making used of the more optimized
version in Graphs.jl, I think the extra abstraction also gives an
insight into what exactly the induced orientation of the graph is
that we're using to find strongly connected components.

While we're at it also replace the hardcoded integer sentintel by
a singleton type to align more with how we're doing this elsewhere
in Julia.