Refactor tearing code
This refactors tearing to use a clean implementation of the BFGT15 Algorithm N,
together with the directed graph abstraction introduced in #1328. I believe
the asymptotics of the version I implemented in Graphs.jl are somewhat better
than the version we're replacing here from Modia, because the Modia versions
caches the level rather than recomputing it (thus potential performing
redundant updates) - however, this is probably unlikely to make a difference
in practice. That said, this version should also be significantly easier to
swap out for one of the more advanced cycle detection algorithms from the
literature. Algorithm N is tuned for dense graphs, which our MTK graphs are
unlikely to be, so we may want to switch to a sparse cycle detection algorithm
in the future.