Refactor (inv_)assign/{var/eq}assoc variables into proper structs
These various assignments and associations look very similar, but
I can never remember which one goes in which direction and which one
is a matching and which one is a derivative ladder. Make this all
more clear by factoring into to types `Matching` for matching and
`DiffGraph` for derivative ladders. Both have a vector structure
and the latter also exposes a graph structure on the same vertex
set as the BipartiteGraph, which is useful for plotting (not part
of this PR). I've also renamed the variables to make the direction
clear.