Strengthen alias elimination
Previously, we would only alias eliminate variables that do not occur
differentiated. This was unnecessarily restricted for systems such as
```
x ~ 0
D(x) ~ x - y
```
where alias elimination is now able to determine that `x` and `y` are
aliases.

While this is the main point of the PR, this does improve alias elimination
in some of the tests, causing tearing order to change. Previously we had
a suboptimality in tearing where an equation assigned to an scc in which it
is not able to solve a variable would not be reconsidered in another scc.
Fix that also, to make sure the tests keep working.