Refactor to pull graph manipulation out of patenlides
It seemes a bit weird to update the graph itself in pantelides, but
update the solvable_graph in eq_derivative!, since it's not clear who
is in charge of the data structure. Refactor this to put eq_derivative!
in charge of updating both data structures, but pull out the graph
manipulation code into a separate helper function that can be used
by non-symbolic eq_derivative implementations to avoid repeating code.