Fix accidental aliasing between mm and structure.graph
After the set_neighbors! call, future modifications to the graph
would silently corrupt the coefficient matrix. We didn't notice
because we're currently not doing that, but it's a big footgun
that I hit trying to do some fancier things, so fix this up
in the hope to prevent others from hitting it in the future.