Minor refactor: Be consistent about alias_eliminate_graph! return type
Generally, it's bad style to change the return type in a corner cases,
because a lot of the time (as I happed to), users will mostly test
the common case and then when the corner case (in this case no linear
equations in the system) comes up, things break. Simplify things
by being consistent about the return types.