Add verbose REPL printing for SystemStructure
This adds REPL printing for SystemStructure that prints out the details
of graph, solvable_graph, var_to_diff and eq_to_diff in a compact, but
comprehensive dump. Because this is a lot of information, it is not
provided by default. Instead, the user may opt into it by setting `:limit`
or `:mtk_limit` to false in the IOContext. E.g. to enable the verbose
printing for all structures shown in the REPL, use

```
Base.active_repl.options.iocontext[:mtk_limit] = false
```