state_selection: Correct is_not_present condition
The is_present condition is supposed to be equivalent to the variable
not appearing in the system at all (for example, it could appear as
`a*D(v)`, where `a` ends up aliased to `0`, meaning that we don't
notice that it does not appear in the system until we've performed
alias analysis. In these kinds of situations, a variable is present
in the system if, either:

1. It appears in the graph (i.e. it's used in the equations), or
2. It is irreducible (and was in the alias graph).

The condition was trying to check this, but didn't quite get the
negations correct, resulting in sub-optimal state selection.