dummy_derivatives: Reset eqcolor every loop
We're using augmenting path construction here to try to find a maximal
matching of the variables at the given differentiation level. However
the construct_augmenting_path! function assumes as a pre-condition that
eq-color is reset to zero in order for it to actually produce an
augmenting path. Failing to do this can cause it to falsely declare
a system singular.