avoid n^2 scaling in expand_instream
Before this patch, the connection set that belongs to an instream expression was found by looping over all of them. With n expressions and n connection sets, the time this took for large systems became very large (at least a day for me). This fixes that by creating a Dict that maps directly to the required connection set. This Dict is created by looping over all connection sets once.