Fix Barreiss algorithm (#2089)
This has a number of changes.
The first is that we were using ÷ instead of exact_div in bareiss_update_virtual_colswap_mtk
which led to undetected overflows. We also had a number of bugs that arise when using BigInt precsion in bareiss (e.g. checking `!==0` vs `!=0`.
This PR makes it so we first try to bareiss with the regular Int64 matrix, but if that gets an overflow, we fallback to
the BigInt version.