Change `show(::MatchedSystemStructure)` to use fewer colors.
This includes the following changes for `Base.show(::IO,
::MatchedSystemStructure)`:
  - var ⇔ eq matches are marked with underline (previously, yellow)
  - (un)solvable vars/eqs are printed in dark gray and white,
    respectively (previously, white/yellow)
  - the ∫ symbol is now printed without coloring
  - A separator has been added for the var/eq sides of the table

These changes are made to encode less semantic information using
colors, so that can colors instead be used to present a unified
"diff" view of these objects. The new format is arguably more
boring, but IMO is also easier to scan for information.