Add case for when reduced `mm` row containing only single entry
Because `mm` represents a homogenous linear system, if there is a
single entry in a row of `mm`, that implicitly states that the entry
must be zero.  This commit enforces this by forcing the corresponding
entries of `ag` to be zero.

Because this only known case where this appears is in a rather involved
reproducer (using DAECompiler to run ModelingToolkitStandardLibrary
tests) we extract the relevant graphs and feed it into MTK here to act
as a test case.