tearing: Heuristically attempt to assign equations with single solvable first
This changes to tearing to give priority to any equation <-> variable pairs
where a particular equation only has one matchable variable. These are easy,
becuse, if the equation is part of the final matching at all, we know which
variable they need to be matched to and by matching them first, we can avoid
accidentally assigning these variables to another equation, which would
necessarily increase the number of torn states.