Handle dummy derivatives in `linearization_function`
The issue here is that the homogenous parameter PR that was merged broke `linearization_function`, which led to a further PR that broke the user interface to `linearization_function` such that the user now needs to provide an "example input" that encodes the types that the generated function would eventually be called with.

I find this a very bad design and a quite unfortunate breaking change. The fundamental problem is that state and parameter types are important, but the modeling language does not allow the user to encode this information in the model, instead we now ask the users to provide the type information in the form of an example input.