Correct number of srcs in `complete`
When constructing the inverse match, `complete` was allocating a vector
with length of the forward match. This isn't quite the right number.
What you want is to use the actual number of srcs (or failing that
the maximum number that appears in the matching). In practice, this
was mostly working out, because an overestimate is fine here and
usually there are more variables than equations, but there don't
have to be in all places where Matching is used. Fix that by defaulting
`complete` to the maximum used src and explicitly passing the correct
number of sources in a few places.