bariess: Don't try to process pivot variable in bareiss update
At the top of the bareiss_update_virtual_colswap_mtk! function,
the first thing we do is to delete the pivot column from the
substitution row (so the new coefficient is zero). However,
we were still revisiting this row in the update loop (but
throwing away the answer). This could trigger an assertion,
because a zero in that position is not guaranteed to satisfy
the fraction-freeness criterion of the bareiss algorithm.
Fix that by bailing out earlier when the variable we encounter
is the pivot.