Fix an exception during documentation build
Referencing the non-existent `SciMLBase.SciMLProblem` causes an
exception during documentation build because `SiMLBase` is not in scope.
This exception is caught and doesn't break the build, but it does result
in a missing link in the documentation. Adding an import for `SCIMLBase`
and referencing the correct `SciMLBase.AbstractSciMLProblem` fixes the
error.