Remove incidence matrix from system show
Fixes https://github.com/SciML/ModelingToolkit.jl/issues/2178

>
The incidence matrix is actually not useful for humans. I am just showing it because it looks cool. We can just disable it if it's confusing.

I'm not even sure it's worth documenting, but who it's useful for is for the compiler writers since if structural simplification works, then you're good. Maybe it's useful for diagnosing singularities, but that should get a debugging page because someone would need to get introduced to what it is for it to be useful. So I keep it around as a `ModelingToolkit.incidence_matrix` function that can be easily queried, but it's removed from the default show.