Consider non-timeseries parameters to be ContinuousTimeseries
This does a few things. First of all it fixes https://github.com/SciML/SciMLBase.jl/issues/1061 by making parameters plot as lines. But secondly, it treats parameters the same way as variables, which is a requirement if we want to make it simpler for the compiler to change things from the state to parameters.