add test for passing params as structural params
it is common to define composite systems, where the parameters and their
names are defined in the "outer" system. Passing them as structural
parameters ensures, that there is only one parameter holding that
information (which can be changed with a callback for example)
rather than having multiple parameters which just share the same
default values.